+++ src/components/Main/Main.js
@@ -5,12 +5,12 @@ import VideoSection from 'components/Video'
 import './Main.scss'
 import MainSide from './MainSide'
 
-const Main = ({ InfoMenus }) => {
+const Main = ({ InfoMenus, changeMenu, menuIndex }) => {
   return (
     <div className={'mainpage'}>
       <div className={'main_container'}>
         <VideoSection />
-        <InfoLayout menus={InfoMenus} />
+        <InfoLayout menus={InfoMenus} menuIndex={menuIndex} changeMenu={changeMenu} />
       </div>
       <MainSide />
       <img src={backgroundImg} className={'background'} alt={'background'} />
